% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make_groups}
\usage{
make_groups(names)
}
\arguments{
\item{names}{}
}
\value{
A two column matrix of names
}
\description{
Makes groups of two, randomly, from a list of names.
Has a safeguard by which it stops is there is an
Uneven number of people in the list
}
\examples{
x

}
