% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutte_p2_edgelist.R
\docType{data}
\name{rutte_p2_edgelist}
\alias{rutte_p2_edgelist}
\title{rutte_p2_edgelist}
\format{
An object of class \code{data.frame} with 106 rows and 10 columns.
}
\source{
: https://www-tandfonline-com.proxy.library.uu.nl/doi/full/10.1080/13501763.2022.2155215
}
\usage{
rutte_p2_edgelist
}
\arguments{
\item{rutte_p2_edgelist}{}
}
\value{
a dataframe with the data of the cognitive map of the Dutch Prime Minister
Mark Rutte from speeches between May 2010 and July 2012 which is used in
Van Esch & Snellens, online first


The df contains the following data:
from: id of the source/cause concept
to: id of the target/effect concept
weight:
edge_value:
edge_id:
map_id:
map-date:
speaker:
value.x:
value.y:
}
\description{
rutte_p2_edgelist
}
\keyword{datasets}
